+++ internal/dbserver/repository/task_repository.go
@@ -17,7 +17,7 @@ type TaskRepository interface {
 	GetTaskDataByID(taskID int64) (types.TaskData, error)
 	GetTasksByJobID(jobID int64) ([]types.TasksByJobIDResponse, error)
 	UpdateTaskFee(taskID int64, fee float64) error
-	GetTaskFee(taskID int64) (int64, error)
+	GetTaskFee(taskID int64) (float64, error)
 }
 
 type taskRepository struct {
@@ -109,8 +109,8 @@ func (r *taskRepository) UpdateTaskFee(taskID int64, fee float64) error {
 	return nil
 }
 
-func (r *taskRepository) GetTaskFee(taskID int64) (int64, error) {
-	var fee int64
+func (r *taskRepository) GetTaskFee(taskID int64) (float64, error) {
+	var fee float64
 	err := r.db.Session().Query(queries.GetTaskFeeQuery, taskID).Scan(&fee)
 	if err != nil {
 		return 0, errors.New("error getting task fee")
